Title: 피보나치 수, Time: 0ms, Memory: 0MB, Status: Runtime Error - Codetree

+++ 260406/피보나치 수/fibonacci-number.java
@@ -6,7 +6,7 @@ public class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		
 		int N = Integer.parseInt(br.readLine());
-		int[] dp = new int[N+1];
+		long[] dp = new long[N+1];
 		
 		dp[1] = 1;
 		dp[2] = 1;
